+++ sdk/keyvault/azure-keyvault-keys/tests/test_key_client.py
@@ -13,22 +13,26 @@ import sys
 import json
 
 from azure.core.exceptions import ResourceNotFoundError
-from azure.keyvault.keys import JsonWebKey
-from keys_preparer import VaultClientPreparer
-from keys_test_case import KeyVaultTestCase
+from azure.keyvault.keys import JsonWebKey, KeyClient
 from devtools_testutils import ResourceGroupPreparer, KeyVaultPreparer
 
+from _shared.preparer import KeyVaultClientPreparer as _KeyVaultClientPreparer
+from _shared.test_case import KeyVaultTestCase
+
+# pre-apply the client_cls positional argument so it needn't be explicitly passed below
+KeyVaultClientPreparer = functools.partial(_KeyVaultClientPreparer, KeyClient)
 
 # used for logging tests
 class MockHandler(logging.Handler):
     def __init__(self):
         super(MockHandler, self).__init__()
         self.messages = []
+
     def emit(self, record):
         self.messages.append(record)
 
-class KeyClientTests(KeyVaultTestCase):
 
+class KeyClientTests(KeyVaultTestCase):
     def _assert_key_attributes_equal(self, k1, k2):
         self.assertEqual(k1.name, k2.name)
         self.assertEqual(k1.vault_url, k2.vault_url)
@@ -45,7 +49,9 @@ class KeyClientTests(KeyVaultTestCase):
         key_size = 2048
         key_ops = ["encrypt", "decrypt", "sign", "verify", "wrapKey", "unwrapKey"]
         tags = {"purpose": "unit test", "test name ": "CreateRSAKeyTest"}
-        created_key = client.create_rsa_key(key_name, hardware_protected=hsm, size=key_size, key_operations=key_ops, tags=tags)
+        created_key = client.create_rsa_key(
+            key_name, hardware_protected=hsm, size=key_size, key_operations=key_ops, tags=tags
+        )
         self.assertTrue(created_key.properties.tags, "Missing the optional key attributes.")
         self.assertEqual(tags, created_key.properties.tags)
         kty = "RSA-HSM" if hsm else "RSA"
@@ -73,7 +79,8 @@ class KeyClientTests(KeyVaultTestCase):
         self.assertTrue(kid.index(prefix) == 0, "Key Id should start with '{}', but value is '{}'".format(prefix, kid))
         self.assertEqual(key.kty, kty, "kty should by '{}', but is '{}'".format(key, key.kty))
         self.assertTrue(
-            key_attributes.properties.created_on and key_attributes.properties.updated_on, "Missing required date attributes."
+            key_attributes.properties.created_on and key_attributes.properties.updated_on,
+            "Missing required date attributes.",
         )
 
     def _validate_rsa_key_bundle(self, key_attributes, vault, key_name, kty, key_ops):
@@ -85,7 +92,8 @@ class KeyClientTests(KeyVaultTestCase):
         self.assertTrue(key.n and key.e, "Bad RSA public material.")
         self.assertEqual(key_ops, key.key_ops, "keyOps should be '{}', but is '{}'".format(key_ops, key.key_ops))
         self.assertTrue(
-            key_attributes.properties.created_on and key_attributes.properties.updated_on, "Missing required date attributes."
+            key_attributes.properties.created_on and key_attributes.properties.updated_on,
+            "Missing required date attributes.",
         )
 
     def _update_key_properties(self, client, key):
@@ -137,11 +145,10 @@ class KeyClientTests(KeyVaultTestCase):
 
     @ResourceGroupPreparer(random_name_enabled=True)
     @KeyVaultPreparer()
-    @VaultClientPreparer()
-    def test_key_crud_operations(self, vault_client, **kwargs):
+    @KeyVaultClientPreparer()
+    def test_key_crud_operations(self, client, **kwargs):
 
-        self.assertIsNotNone(vault_client)
-        client = vault_client.keys
+        self.assertIsNotNone(client)
 
         # create ec key
         self._create_ec_key(client, key_name="crud-ec-key", hsm=True)
@@ -188,11 +195,10 @@ class KeyClientTests(KeyVaultTestCase):
 
     @ResourceGroupPreparer(random_name_enabled=True)
     @KeyVaultPreparer(enable_soft_delete=False)
-    @VaultClientPreparer()
-    def test_backup_restore(self, vault_client, **kwargs):
+    @KeyVaultClientPreparer()
+    def test_backup_restore(self, client, **kwargs):
 
-        self.assertIsNotNone(vault_client)
-        client = vault_client.keys
+        self.assertIsNotNone(client)
         key_name = self.get_resource_name("keybak")
         key_type = "RSA"
 
@@ -213,11 +219,10 @@ class KeyClientTests(KeyVaultTestCase):
 
     @ResourceGroupPreparer(random_name_enabled=True)
     @KeyVaultPreparer()
-    @VaultClientPreparer()
-    def test_key_list(self, vault_client, **kwargs):
+    @KeyVaultClientPreparer()
+    def test_key_list(self, client, **kwargs):
 
-        self.assertIsNotNone(vault_client)
-        client = vault_client.keys
+        self.assertIsNotNone(client)
 
         max_keys = self.list_test_size
         expected = {}
@@ -238,11 +243,10 @@ class KeyClientTests(KeyVaultTestCase):
 
     @ResourceGroupPreparer(random_name_enabled=True)
     @KeyVaultPreparer()
-    @VaultClientPreparer()
-    def test_list_versions(self, vault_client, **kwargs):
+    @KeyVaultClientPreparer()
+    def test_list_versions(self, client, **kwargs):
 
-        self.assertIsNotNone(vault_client)
-        client = vault_client.keys
+        self.assertIsNotNone(client)
         key_name = self.get_resource_name("testKey")
 
         max_keys = self.list_test_size
@@ -265,10 +269,9 @@ class KeyClientTests(KeyVaultTestCase):
 
     @ResourceGroupPreparer(random_name_enabled=True)
     @KeyVaultPreparer()
-    @VaultClientPreparer()
-    def test_list_deleted_keys(self, vault_client, **kwargs):
-        self.assertIsNotNone(vault_client)
-        client = vault_client.keys
+    @KeyVaultClientPreparer()
+    def test_list_deleted_keys(self, client, **kwargs):
+        self.assertIsNotNone(client)
 
         expected = {}
 
@@ -297,10 +300,9 @@ class KeyClientTests(KeyVaultTestCase):
 
     @ResourceGroupPreparer(random_name_enabled=True)
     @KeyVaultPreparer()
-    @VaultClientPreparer()
-    def test_recover(self, vault_client, **kwargs):
-        self.assertIsNotNone(vault_client)
-        client = vault_client.keys
+    @KeyVaultClientPreparer()
+    def test_recover(self, client, **kwargs):
+        self.assertIsNotNone(client)
 
         # create keys
         keys = {}
@@ -324,10 +326,9 @@ class KeyClientTests(KeyVaultTestCase):
 
     @ResourceGroupPreparer(random_name_enabled=True)
     @KeyVaultPreparer()
-    @VaultClientPreparer()
-    def test_purge(self, vault_client, **kwargs):
-        self.assertIsNotNone(vault_client)
-        client = vault_client.keys
+    @KeyVaultClientPreparer()
+    def test_purge(self, client, **kwargs):
+        self.assertIsNotNone(client)
 
         # create keys
         key_names = ["key{}".format(i) for i in range(self.list_test_size)]
@@ -356,22 +357,21 @@ class KeyClientTests(KeyVaultTestCase):
 
     @ResourceGroupPreparer(random_name_enabled=True)
     @KeyVaultPreparer()
-    @VaultClientPreparer(client_kwargs={'logging_enable': True})
-    def test_logging_enabled(self, vault_client, **kwargs):
-        client = vault_client.keys
+    @KeyVaultClientPreparer(client_kwargs={"logging_enable": True})
+    def test_logging_enabled(self, client, **kwargs):
         mock_handler = MockHandler()
 
-        logger = logging.getLogger('azure')
+        logger = logging.getLogger("azure")
         logger.addHandler(mock_handler)
         logger.setLevel(logging.DEBUG)
 
         client.create_rsa_key("rsa-key-name", size=2048)
 
         for message in mock_handler.messages:
-            if message.levelname == 'DEBUG' and message.funcName == 'on_request':
+            if message.levelname == "DEBUG" and message.funcName == "on_request":
                 try:
                     body = json.loads(message.message)
-                    if body['kty'] == 'RSA':
+                    if body["kty"] == "RSA":
                         return
                 except (ValueError, KeyError):
                     # this means the message is not JSON or has no kty property
@@ -381,19 +381,18 @@ class KeyClientTests(KeyVaultTestCase):
 
     @ResourceGroupPreparer(random_name_enabled=True)
     @KeyVaultPreparer()
-    @VaultClientPreparer()
-    def test_logging_disabled(self, vault_client, **kwargs):
-        client = vault_client.keys
+    @KeyVaultClientPreparer()
+    def test_logging_disabled(self, client, **kwargs):
         mock_handler = MockHandler()
 
-        logger = logging.getLogger('azure')
+        logger = logging.getLogger("azure")
         logger.addHandler(mock_handler)
         logger.setLevel(logging.DEBUG)
 
         client.create_rsa_key("rsa-key-name", size=2048)
 
         for message in mock_handler.messages:
-            if message.levelname == 'DEBUG' and message.funcName == 'on_request':
+            if message.levelname == "DEBUG" and message.funcName == "on_request":
                 try:
                     body = json.loads(message.message)
                     assert body["kty"] != "RSA", "Client request body was logged"