+++ sdk/core/azure-core/CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## 1.2.3 (Unreleased)
 
+### Bug fixes
+
+- Appended RequestIdPolicy to the default pipeline
 
 ## 1.2.2 (2020-02-10)